serverDefinitions servers mapped instead of string
Previously the values.yaml file had to contain the serverDefinitions as
followed:

```
serverDefinitions:
  servers: |-
    "1": {
      "Name": "Minimally Defined Server",
      "Group": "Servers",
      ...
    }
```

With this change, the following data structure is used:

```
serverDefinitions:
  servers:
    firstServer:
      Name: "Minimally Defined Server"
      Group: "Servers"
      ...
```

Be aware that this can break your current release, if upgrading to this
helm chart version 1.9.0.

This solution was brought to my attention by https://github.com/itscaro,
so thanks!

Signed-off-by: Rowan Ruseler <rowanruseler@gmail.com>